Mdbport

+++ build.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+source .env
+echo "Using conn string starting ${MDBCONNSTR:0:18}..."
+
+datehash=`date | md5sum | cut -d" " -f1`
+abbrvhash=${datehash: -8}
+
 echo
 echo "+======================"
 echo "| START: Heres Greg"
@@ -7,26 +13,33 @@ echo "+======================"
 echo
 
 echo 
-echo "GREG: Building webapp"
+echo "Building container using tag ${abbrvhash}"
 echo
-cd HeresKids
-dotnet clean
-cd ..
-echo 
-echo "GREG: Building container"
-echo
-docker build -t graboskyc/heresgregblazor:latest .
+docker build -t graboskyc/heresgregblazor:${abbrvhash} .
+
+EXITCODE=$?
 
 echo 
 echo "GREG: Starting container"
 echo
 
-docker stop heresgregblazor
-docker rm heresgregblazor
-docker run -t -i -d -p 9999:80 --name heresgregblazor --restart unless-stopped graboskyc/heresgregblazor:latest
+if [ $EXITCODE -eq 0 ]
+    then
 
-echo
-echo "+======================"
-echo "| END: Heres Greg"
-echo "+======================"
-echo
+    docker stop heresgregblazor
+    docker rm heresgregblazor
+    docker run -t -i -d -p 9999:8080 --name heresgregblazor -e "MDBCONNSTR=${MDBCONNSTR}" -e "HOSTEDURI=${HOSTEDURI}" -e "JWTKEY=${JWTKEY}" -e "DEPLOYMENTBASEURI=${DEPLOYMENTBASEURI}" --restart unless-stopped graboskyc/heresgregblazor:${abbrvhash}
+
+    echo
+    echo "+================================"
+    echo "| END: Heres Greg"
+    echo "+================================"
+    echo
+
+else
+    echo
+    echo "+================================"
+    echo "| ERROR: Build failed"
+    echo "+================================"
+    echo
+fi
